requirements.txt

Flask
gTTS
SpeechRecognition
python-docx
pydub
